requirements.txt
langchain
langchain_community
chromadb
tiktoken
chainlit
openai
python-dotenv